PlatformMapper.javaとPlatformMapper.xmlを作成
