+++ src/test/java/com/crud_read_and_create/service/GameServiceTest.java
@@ -5,6 +5,7 @@ import com.crud_read_and_create.entity.Game;
 import com.crud_read_and_create.entity.GamePlatform;
 import com.crud_read_and_create.entity.Platform;
 import com.crud_read_and_create.mapper.GameMapper;
+import com.crud_read_and_create.mapper.PlatformMapper;
 import com.crud_read_and_create.service.exception.DuplicateException;
 import com.crud_read_and_create.service.exception.NotFoundException;
 import org.junit.jupiter.api.Test;
@@ -29,6 +30,9 @@ class GameServiceTest {
 
     @MockBean(name = "gameMapper")
     private GameMapper gameMapper;
+
+    @MockBean(name = "platformMapper")
+    private PlatformMapper platformMapper;
 
     @Autowired
     private GameService gameService;
@@ -70,7 +74,7 @@ class GameServiceTest {
     @Test
     void プラットフォームを全件取得できること() {
         List<Platform> platformList = Arrays.asList(new Platform(1, "PS5"), new Platform(2, "Switch"));
-        doReturn(platformList).when(gameMapper).findPlatform();
+        doReturn(platformList).when(platformMapper).findPlatform();
 
         List<Platform> actualPlatformList = gameService.getPlatform();
         assertThat(actualPlatformList).hasSize(2).isEqualTo(platformList);
@@ -103,7 +107,7 @@ class GameServiceTest {
         String[] platformIds = {"1", "2"};
         List<GamePlatform> gamePlatformList = Arrays.asList(new GamePlatform(game.getId(), platformIds[0]),
                 new GamePlatform(game.getId(), platformIds[1]));
-        
+
         gameService.createGame(game.getId(), game.getName(), game.getGenre(), game.getPrice(), platformIds);
         verify(gameMapper, times(1)).createGame(gameCapture.capture());
         verify(gameMapper, times(1)).createGamePlatform(gamePlatformCapture.capture());
@@ -116,7 +120,7 @@ class GameServiceTest {
     @Test
     void 入力したプラットフォームがすでに登録されている時_DuplicateExceptionが発生すること() {
         List<Platform> platformList = Arrays.asList(new Platform(1, "PS5"), new Platform(2, "Switch"));
-        doReturn(platformList).when(gameMapper).findPlatform();
+        doReturn(platformList).when(platformMapper).findPlatform();
         assertThrows(DuplicateException.class, () -> gameService.createPlatform(1, "PS5"));
         assertThatThrownBy(() -> {
             throw new DuplicateException("プラットフォームが重複しています。");
@@ -130,11 +134,11 @@ class GameServiceTest {
     @Test
     void 入力したプラットフォームが未登録の時_引数にしていたPlatformが正常に登録処理が実行されること() throws DuplicateException {
         List<Platform> platformList = Arrays.asList(new Platform(1, "PS5"), new Platform(2, "Switch"));
-        doReturn(platformList).when(gameMapper).findPlatform();
+        doReturn(platformList).when(platformMapper).findPlatform();
         Platform platformData = new Platform(3, "Steam");
 
         gameService.createPlatform(platformData.getId(), platformData.getPlatform());
-        verify(gameMapper, times(1)).createPlatform(platformCapture.capture());
+        verify(platformMapper, times(1)).createPlatform(platformCapture.capture());
         Platform actualPlatform = platformCapture.getValue();
         assertThat(actualPlatform).isEqualTo(platformData);
     }
